refactor(PlaceMap): View 시스템 코드 제거 및 패키지 구조 통합
장소 지도(`PlaceMap`) 화면의 Compose 마이그레이션이 완료됨에 따라, 더 이상 사용하지 않는 기존 Fragment 및 XML 레이아웃 코드를 삭제했습니다. 또한, 기능별로 분산되어 있던 Compose 컴포넌트와 관련 클래스들을 `placeMap` 패키지 하위로 통합하여 구조를 단순화했습니다.

- **Legacy 코드 삭제:**
    - `PlaceListFragment`, `PlaceCategoryFragment`, `PlaceDetailPreviewFragment` 등 기존 View 기반의 Fragment 클래스들을 삭제했습니다.
    - `fragment_place_list.xml`, `item_place_list.xml` 등 관련 XML 레이아웃 리소스와 바인딩 코드를 제거했습니다.

- **패키지 구조 재편:**
    - `placeList`, `placeCategory`, `placeDetailPreview` 등 하위 패키지에 흩어져 있던 컴포저블(`PlaceListScreen`, `PlaceCategoryScreen` 등)과 `PlaceListViewModel`을 `presentation.placeMap` 및 `presentation.placeMap.component` 패키지로 이동하여 접근성을 높였습니다.
    - `PlaceMapFragment` 및 `PlaceMapScreen`에서 변경된 패키지 경로를 반영하도록 import 구문을 수정했습니다.

- **`CategoryView` 이동:**
    - 장소 상세 화면(`PlaceDetailActivity`)에서 여전히 사용되는 커스텀 뷰 `CategoryView`를 `presentation.placeDetail` 패키지로 이동하고, `activity_place_detail.xml`에서 해당 뷰를 참조하도록 수정했습니다.
