+++ app/src/main/java/com/daedan/festabook/di/mapManager/MapManagerBindings.kt
@@ -1,11 +1,11 @@
 package com.daedan.festabook.di.mapManager
 
-import com.daedan.festabook.presentation.placeMap.MapClickListener
-import com.daedan.festabook.presentation.placeMap.MapClickListenerImpl
-import com.daedan.festabook.presentation.placeMap.PlaceMapViewModel
+import com.daedan.festabook.presentation.placeMap.listener.MapClickListener
+import com.daedan.festabook.presentation.placeMap.listener.MapClickListenerImpl
 import com.daedan.festabook.presentation.placeMap.mapManager.internal.OverlayImageManager
 import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
 import com.daedan.festabook.presentation.placeMap.model.iconResources
+import com.daedan.festabook.presentation.placeMap.viewmodel.PlaceMapViewModel
 import com.naver.maps.map.overlay.Marker
 import dev.zacsweers.metro.BindingContainer
 import dev.zacsweers.metro.ContributesTo
