+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/component/OffsetDependentLayout.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeMap.placeList.component
+package com.daedan.festabook.presentation.placeMap.component
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
