+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/component/PlaceListBottomSheetState.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeMap.placeList.component
+package com.daedan.festabook.presentation.placeMap.component
 
 import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.Spring
