+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/listener/MapClickListener.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeMap
+package com.daedan.festabook.presentation.placeMap.listener
 
 import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
