+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/listener/OnCameraChangeListener.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeMap
+package com.daedan.festabook.presentation.placeMap.listener
 
 fun interface OnCameraChangeListener {
     fun onCameraChanged(isExceededMaxLength: Boolean)
