+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/mapManager/internal/MapCameraManagerImpl.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.presentation.placeMap.mapManager.internal
 
 import com.daedan.festabook.di.mapManager.PlaceMapScope
-import com.daedan.festabook.presentation.placeMap.OnCameraChangeListener
+import com.daedan.festabook.presentation.placeMap.listener.OnCameraChangeListener
 import com.daedan.festabook.presentation.placeMap.mapManager.MapCameraManager
 import com.daedan.festabook.presentation.placeMap.model.InitialMapSettingUiModel
 import com.daedan.festabook.presentation.placeMap.model.toLatLng
