--- app/src/main/java/com/daedan/festabook/presentation/placeMap/placeList/OnPlaceClickListener.kt
@@ -1,7 +1,0 @@
-package com.daedan.festabook.presentation.placeMap.placeList
-
-import com.daedan.festabook.presentation.placeMap.model.PlaceUiModel
-
-fun interface OnPlaceClickListener {
-    fun onPlaceClicked(place: PlaceUiModel)
-}
