+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/viewmodel/PlaceMapAction.kt
@@ -1,0 +1,3 @@
+package com.daedan.festabook.presentation.placeMap.viewmodel
+
+class PlaceMapAction
