+++ app/src/main/java/com/daedan/festabook/presentation/placeMap/viewmodel/PlaceMapViewModel.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeMap
+package com.daedan.festabook.presentation.placeMap.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
