--- app/src/main/res/layout/fragment_place_detail_preview.xml
@@ -1,152 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_marginHorizontal="16dp"
-        android:layout_width="match_parent"
-        android:layout_gravity="bottom|center"
-        android:layout_marginBottom="8dp"
-        android:layout_height="wrap_content">
-
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/layout_selected_place"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@drawable/bg_gray050_stroke_gray200_radius_24dp"
-            android:paddingVertical="20dp"
-            android:paddingHorizontal="16dp">
-
-            <com.daedan.festabook.presentation.placeMap.placeList.CategoryView
-                android:id="@+id/cv_place_category"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="8dp"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintStart_toStartOf="parent" />
-
-            <TextView
-                android:id="@+id/tv_selected_place_title"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:textAppearance="@style/PretendardBold18"
-                android:textColor="@color/gray900"
-                android:transitionName="selected_place_title_transition"
-                android:layout_marginTop="4dp"
-                android:maxLines="2"
-                android:ellipsize="end"
-                android:layout_marginEnd="4dp"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toStartOf="@id/cv_place_image"
-                app:layout_constraintTop_toBottomOf="@id/cv_place_category"
-                tools:text="컴공과 주점 코딩하며 한잔컴공과 주점 코딩하며 한잔" />
-
-            <ImageView
-                android:id="@+id/iv_selected_place_time"
-                android:layout_width="16dp"
-                android:layout_height="16dp"
-                android:layout_marginTop="12dp"
-                android:contentDescription="운영 시간"
-                android:src="@drawable/ic_place_detail_clock"
-                app:layout_constraintStart_toStartOf="@id/tv_selected_place_title"
-                app:layout_constraintTop_toBottomOf="@id/tv_selected_place_title" />
-
-            <TextView
-                android:id="@+id/tv_selected_place_time"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="4dp"
-                android:textAppearance="@style/PretendardRegular12"
-                android:textColor="@color/gray500"
-                app:layout_constraintBottom_toBottomOf="@id/iv_selected_place_time"
-                app:layout_constraintStart_toEndOf="@id/iv_selected_place_time"
-                app:layout_constraintTop_toTopOf="@id/iv_selected_place_time"
-                tools:text="09:00 ~ 18:00" />
-
-            <ImageView
-                android:id="@+id/iv_selected_place_location"
-                android:layout_width="16dp"
-                android:layout_height="16dp"
-                android:layout_marginTop="4dp"
-                android:contentDescription="위치"
-                android:src="@drawable/ic_location"
-                app:layout_constraintStart_toStartOf="@id/iv_selected_place_time"
-                app:layout_constraintTop_toBottomOf="@id/iv_selected_place_time" />
-
-            <TextView
-                android:id="@+id/tv_selected_place_location"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="4dp"
-                android:textAppearance="@style/PretendardRegular12"
-                android:textColor="@color/gray500"
-                app:layout_constraintBottom_toBottomOf="@id/iv_selected_place_location"
-                app:layout_constraintStart_toEndOf="@id/iv_selected_place_location"
-                app:layout_constraintTop_toTopOf="@id/iv_selected_place_location"
-                tools:text="공학관 앞" />
-
-            <ImageView
-                android:id="@+id/iv_selected_place_host"
-                android:layout_width="16dp"
-                android:layout_height="16dp"
-                android:layout_marginTop="4dp"
-                android:contentDescription="운영자"
-                android:src="@drawable/ic_place_detail_host"
-                app:layout_constraintStart_toStartOf="@id/iv_selected_place_location"
-                app:layout_constraintTop_toBottomOf="@id/iv_selected_place_location" />
-
-            <TextView
-                android:id="@+id/tv_selected_place_host"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="4dp"
-                android:textAppearance="@style/PretendardRegular12"
-                android:textColor="@color/gray500"
-                app:layout_constraintBottom_toBottomOf="@id/iv_selected_place_host"
-                app:layout_constraintStart_toEndOf="@id/iv_selected_place_host"
-                app:layout_constraintTop_toTopOf="@id/iv_selected_place_host"
-                tools:text="C블C블" />
-
-            <TextView
-                android:id="@+id/tv_selected_place_description"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="12dp"
-                android:textAppearance="@style/PretendardRegular12"
-                android:autoLink="web"
-                android:textColorLink="@color/gray500"
-                android:maxLines="2"
-                android:ellipsize="end"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="@id/cv_place_image"
-                app:layout_constraintStart_toStartOf="@id/iv_selected_place_host"
-                app:layout_constraintTop_toBottomOf="@id/iv_selected_place_host"
-                tools:text="시원한 맥주와 맛있는 파전" />
-
-            <com.google.android.material.card.MaterialCardView
-                android:id="@+id/cv_place_image"
-                android:layout_width="88dp"
-                android:layout_height="88dp"
-                android:layout_marginTop="8dp"
-                app:cardBackgroundColor="@color/gray300"
-                app:cardCornerRadius="8dp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="@id/tv_selected_place_title"
-                app:strokeWidth="0dp">
-
-                <ImageView
-                    android:id="@+id/iv_selected_place_image"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:contentDescription="@string/content_description_booth_image"
-                    android:scaleType="centerCrop"
-                    app:layout_constraintDimensionRatio="1:1" />
-
-            </com.google.android.material.card.MaterialCardView>
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
-</layout>
