--- app/src/main/res/layout/fragment_place_detail_preview_secondary.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|center"
-        android:layout_marginHorizontal="16dp"
-        android:layout_marginBottom="8dp">
-
-        <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-            android:id="@+id/layout_selected_place"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@drawable/bg_gray100_stroke_gray400_radius_10dp"
-            android:paddingVertical="12dp">
-
-            <ImageView
-                android:id="@+id/iv_secondary_category_item"
-                android:layout_width="24dp"
-                android:layout_height="24dp"
-                android:layout_marginStart="16dp"
-                android:contentDescription="기타플레이스 아이콘"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                tools:src="@drawable/ic_map_category_toilet" />
-
-            <TextView
-                android:id="@+id/tv_selected_place_title"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="8dp"
-                android:layout_marginEnd="16dp"
-                android:textAppearance="@style/PretendardBold18"
-                android:textColor="@color/gray900"
-                android:transitionName="selected_place_title_transition"
-                app:layout_constraintBottom_toBottomOf="@id/iv_secondary_category_item"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@id/iv_secondary_category_item"
-                app:layout_constraintTop_toTopOf="@id/iv_secondary_category_item"
-                tools:text="개쩌는 레전드 화장실" />
-
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
-</layout>
