--- app/src/main/res/layout/fragment_place_list.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/layout_place_list"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/bg_place_list"
-            android:paddingHorizontal="16dp"
-            app:behavior_fitToContents="false"
-            app:behavior_halfExpandedRatio="0.4"
-            app:behavior_peekHeight="70dp"
-            app:recyclerView="@id/rv_places">
-
-            <View
-                android:id="@+id/view_divider"
-                android:layout_width="32dp"
-                android:layout_height="4dp"
-                android:layout_marginTop="12dp"
-                android:background="@drawable/bg_gray400_radius_50dp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-
-            <TextView
-                android:id="@+id/tv_title_place_list"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="32dp"
-                android:text="@string/place_list_title"
-                android:textAppearance="@style/PretendardBold24"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/rv_places"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_marginTop="4dp"
-                android:paddingBottom="80dp"
-                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                app:layout_constraintTop_toBottomOf="@id/tv_title_place_list"
-                tools:itemCount="2"
-                tools:listitem="@layout/item_place_list" />
-
-            <com.facebook.shimmer.ShimmerFrameLayout
-                android:id="@+id/sfl_schedule_skeleton"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:importantForAccessibility="no"
-                android:layout_marginTop="16dp"
-                app:layout_constraintTop_toBottomOf="@id/tv_title_place_list"
-                app:shimmer_auto_start="true">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="vertical">
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                    <include layout="@layout/item_place_list_skeleton" />
-
-                </LinearLayout>
-            </com.facebook.shimmer.ShimmerFrameLayout>
-
-            <TextView
-                android:id="@+id/tv_error_to_load_place_info"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="40dp"
-                android:text="@string/empty_state_message"
-                android:textAppearance="@style/PretendardRegular12"
-                android:textColor="@color/gray500"
-                android:visibility="gone"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/rv_places" />
-
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <com.naver.maps.map.widget.LocationButtonView
-            android:id="@+id/lbv_current_location"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="start|bottom"
-            android:layout_marginStart="8dp" />
-
-        <com.google.android.material.chip.Chip
-            android:id="@+id/chip_back_to_initial_position"
-            style="@style/Widget.MaterialComponents.Chip.Action"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center|bottom"
-            android:text="@string/map_back_to_initial_position"
-            android:textAppearance="@style/PretendardBold12"
-            android:textColor="@color/gray900"
-            android:visibility="gone"
-            app:changeVisibility="false"
-            app:chipBackgroundColor="@color/gray050"
-            app:chipStrokeColor="@color/gray500"
-            app:chipStrokeWidth="1dp" />
-    </androidx.coordinatorlayout.widget.CoordinatorLayout>
-</layout>
