--- app/src/main/res/layout/item_place_list.xml
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <data>
-
-        <variable
-            name="place"
-            type="com.daedan.festabook.presentation.placeMap.model.PlaceUiModel" />
-
-        <variable
-            name="listener"
-            type="com.daedan.festabook.presentation.placeMap.placeList.OnPlaceClickListener" />
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:onClick="@{()->listener.onPlaceClicked(place)}"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingBottom="12dp">
-
-        <com.google.android.material.card.MaterialCardView
-            android:id="@+id/cv_place_image"
-            android:layout_width="80dp"
-            android:layout_height="80dp"
-            android:layout_marginTop="8dp"
-            app:cardBackgroundColor="@color/gray300"
-            app:cardCornerRadius="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:strokeWidth="0dp">
-
-            <ImageView
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:contentDescription="@string/content_description_booth_image"
-                android:scaleType="centerCrop"
-                app:imageUrl="@{place.imageUrl}" />
-
-        </com.google.android.material.card.MaterialCardView>
-
-        <com.daedan.festabook.presentation.placeMap.placeList.CategoryView
-            android:id="@+id/cv_place_category"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            app:layout_constraintStart_toEndOf="@id/cv_place_image"
-            app:layout_constraintTop_toTopOf="@id/cv_place_image"
-            app:category="@{place.category}" />
-
-        <TextView
-            android:id="@+id/tv_title_place"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="4dp"
-            android:textColor="@color/gray900"
-            app:layout_constraintStart_toStartOf="@id/cv_place_category"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/cv_place_category"
-            android:textAppearance="@style/PretendardMedium14"
-            android:text="@{place.title ?? @string/place_list_default_title}"
-            tools:text="테스트테스트테스트테스트테스트테스트테스" />
-
-        <TextView
-            android:id="@+id/tv_title_description"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="2dp"
-            android:textColor="@color/gray500"
-            android:maxLines="1"
-            android:ellipsize="end"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="@id/cv_place_category"
-            app:layout_constraintTop_toBottomOf="@id/tv_title_place"
-            android:textAppearance="@style/PretendardRegular12"
-            android:text="@{place.description ?? @string/place_list_default_description}"
-
-            tools:text="시원한 맥주와 맛있는 파전! 시원한 맥주와 맛있는 파전! 시원한 맥주와 맛있는 파전!" />
-
-        <ImageView
-            android:id="@+id/iv_location"
-            android:layout_width="14dp"
-            android:layout_height="14dp"
-            android:layout_marginTop="2dp"
-            android:contentDescription="@string/content_description_iv_location"
-            android:src="@drawable/ic_location"
-            app:layout_constraintBottom_toBottomOf="@id/tv_location"
-            app:layout_constraintStart_toStartOf="@id/cv_place_category"
-            app:layout_constraintTop_toTopOf="@id/tv_location" />
-
-        <TextView
-            android:id="@+id/tv_location"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="1dp"
-            android:layout_marginTop="2dp"
-            android:textColor="@color/gray500"
-            android:ellipsize="end"
-            android:maxLines="1"
-            android:maxEms="10"
-            app:layout_constraintStart_toEndOf="@id/iv_location"
-            app:layout_constraintTop_toBottomOf="@id/tv_title_description"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:textAppearance="@style/PretendardRegular12"
-            android:text="@{place.location ?? @string/place_list_default_location}"
-            tools:text="위치위치위치위치위치위치위치위치" />
-
-        <View
-            android:id="@+id/view_divider"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:layout_marginTop="16dp"
-            android:background="@color/gray300"
-            app:layout_constraintTop_toBottomOf="@id/iv_location" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
